refactor workbench attribution trend route
